requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
requests
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.3.1
drf-yasg>=1.21.7
python-dateutil>=2.8.2
python-dotenv>=1.0.0
# ultralytics  # opzionale: per analisi YOLO telaini

# Modulo Statistiche & AI Analytics
openpyxl==3.1.2
reportlab==4.2.0

